% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sucra.R
\docType{data}
\name{sucra}
\alias{sucra}
\title{SUCRA}
\format{
A data frame with 8 rows and 5 variables:
\describe{
  \item{Outcome.1}{Outcome 1}
  \item{Outcome.5}{Outcome 5}
  \item{Outcome.4}{Outcome 4}
  \item{Outcome.3}{Outcome 3}
  \item{Outcome.2}{Outcome 2}
}
}
\usage{
sucra
}
\description{
A dataset containing example data.
}
\keyword{datasets}
