% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_calculation.R
\name{area_calculation}
\alias{area_calculation}
\title{Function to calculate area of the generated polygon}
\usage{
area_calculation(df)
}
\arguments{
\item{df}{input dataframe in the required format}
}
\value{
result
}
\description{
Function to calculate area of the generated polygon
}
\details{
This function serves as a supplementary tool to compute the area of each generated origami
plot when the maximal area achievable within the defined parameters (when all the variables attain 1) is set to 1.
The resulting calculated area offers an interpretation of the proportion between the actual origami plot and the
maximum achievable area.
}
\examples{
data(sucra)
area_calculation(sucra)

}
