% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sucra.R
\docType{data}
\name{sucra}
\alias{sucra}
\title{SUCRA}
\format{
A data frame with 8 rows and 5 variables:
\describe{
  \item{caesarean}{Caesarean section}
  \item{maternal}{Serious maternal morbidity or death}
  \item{neonatal}{Serious neonatal morbidity or perinatal death}
  \item{hyperstimulation}{Hyperstimulation}
  \item{vaginal}{Vaginal delivery not within 24 hours}
}
}
\usage{
sucra
}
\description{
A dataset containing example data.
}
\keyword{datasets}
